trying t force myself back above 80%

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,12 +6,12 @@
 int main(void)
 {
 	int i;
-	int f1 = 0, f2 = 1, sum;
+	unsigned long f1 = 0, f2 = 1, sum;
 
 	for (i = 0; i < 50; i++)
 	{
 		sum = f1 + f2;
-		printf("%d", sum);
+		printf("%ld", sum);
 
 		f1 = f2;
 		f2 = sum;
